Environmentを1つのファイルに切り出し
